+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Proper Bandersnatch elliptic curve parameters. #TODO
 - `AdminChanged` event parameters no longer indexed. #794
 
 ## [v0.3.0-rc.1] - 2025-08-07